requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.14.0
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
interchange==2021.0.4
monotonic==1.6
packaging==24.1
pansi==2020.7.3
py2neo==2021.2.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
